Start working on collector service implementation.
Creates a new project called nettest-shared-server-storage-couchdb which provides implementations
of the StorageService interfaces. This project is used to abstract the actual database which is
used to store and retrieve measurements. For example, íf PostgreSQL should be the primary database
developers could create a nettest-shared-server-storage-postgresql module and use this instead of
the couchdb module.

The current database model moves from nettest-shared-server project to the new one.
Mapstruct annotation processor is used to map the DTOs to domain objects.

This commit contains a lot added lines because the getter and setter methods of the domain objects
are created.