Add docker-compose configuration for nntool server modules.
This version includes the CouchDB and PostgreSQL databases, the
controller-, collector-, result-service, as well as the ias-server,
qos-service and echo-protocol-server.
The PostgreSQL schema is automatically inserted during container
startup.
For the CouchDB it is currently necessary to manually execute a shell
script to insert the configuration.

The docker-compose configuration is split up into two files. One
contains the general configuration for the containers. The other one
(dev) contains the port definitions and configuration file volumes.
This way, it is easy to switch to a different set of server
configurations by adding a docker-compose.NEW_ENV.yml.