Migrate website to Angular CLI and update to Angular 8.
This commit improves and simplifies the build process by migrating from gulp with
a custom large gulpfile to Angular CLI. Unit and integration (end to end) testing
support is also included.
In the course of this, the Angular version is updated to Angular 8.
NPM now reports zero vulnerabilities after npm install.