Fix potential race condition in QoS task start method.
The field "duration_ns" has not always been set after
the execution of a QoS task. The problem is, that the
operation queue returns from waiting just after the
operation is finished, which happens inside the start method.
Because of that, durationNs variable was sometimes set too
late (after the result has been aquired).
According to documentation from Apple, the start method
of the Operation class should never call super methods.
Therefore, we now override the main method, handle the duration
in this method and call our own "taskMain" method. Now we
can be sure that durationNs will be set prior to finishing
the operation.